Reorganize icon import
