+++ src/components/icon/index.js
@@ -1,15 +1,7 @@
 import React from 'react';
 
-const Icon = ({ height = 16, width = 16, children }) => (
-  <svg
-    width={width}
-    height={height}
-    viewBox={`0 0 ${width} ${height}`}
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    {children}
-  </svg>
-);
+import Add from './add';
+import Close from './close';
+import Update from './update';
 
-export default Icon;
+export { Add, Close, Update };
